+++ app/src/main/java/app/pachli/util/ListStatusAccessibilityDelegate.kt
@@ -1,21 +1,13 @@
 package app.pachli.util
 
-import android.content.ClipData
-import android.content.ClipboardManager
 import android.content.Context
-import android.os.Build
 import android.os.Bundle
 import android.text.Spannable
 import android.text.style.URLSpan
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import android.view.accessibility.AccessibilityManager
-import android.widget.ArrayAdapter
-import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
-import androidx.core.content.ContextCompat.getSystemService
 import androidx.core.view.AccessibilityDelegateCompat
 import androidx.core.view.accessibility.AccessibilityNodeInfoCompat
 import androidx.core.view.accessibility.AccessibilityNodeInfoCompat.AccessibilityActionCompat
@@ -26,7 +18,7 @@ import app.pachli.adapter.FilterableStatusViewHolder
 import app.pachli.adapter.StatusBaseViewHolder
 import app.pachli.core.activity.openLink
 import app.pachli.core.network.model.Status.Companion.MAX_MEDIA_ATTACHMENTS
-import app.pachli.databinding.SimpleListItem1CopyButtonBinding
+import app.pachli.core.ui.ArrayAdapterWithCopyButton
 import app.pachli.interfaces.StatusActionListener
 import app.pachli.viewdata.IStatusViewData
 import app.pachli.viewdata.NotificationViewData
@@ -222,8 +214,9 @@ class ListStatusAccessibilityDelegate<T : IStatusViewData>(
                     ArrayAdapterWithCopyButton(
                         host.context,
                         textLinks,
-                    ),
-                ) { _, which -> host.context.openLink(links[which].link) }
+                    ) { position -> host.context.openLink(links[position].link) },
+                    null,
+                )
                 .show()
                 .let { forceFocus(it.listView) }
         }
@@ -241,10 +234,11 @@ class ListStatusAccessibilityDelegate<T : IStatusViewData>(
                     ArrayAdapterWithCopyButton(
                         host.context,
                         stringMentions,
-                    ),
-                ) { _, which ->
-                    statusActionListener.onViewAccount(mentions[which].id)
-                }
+                    ) { position ->
+                        statusActionListener.onViewAccount(mentions[position].id)
+                    },
+                    null,
+                )
                 .show()
                 .let { forceFocus(it.listView) }
         }
@@ -258,10 +252,11 @@ class ListStatusAccessibilityDelegate<T : IStatusViewData>(
                     ArrayAdapterWithCopyButton(
                         host.context,
                         tags,
-                    ),
-                ) { _, which ->
-                    statusActionListener.onViewTag(tags[which].toString())
-                }
+                    ) { position ->
+                        statusActionListener.onViewTag(tags[position].toString())
+                    },
+                    null,
+                )
                 .show()
                 .let { forceFocus(it.listView) }
         }
@@ -412,41 +407,4 @@ class ListStatusAccessibilityDelegate<T : IStatusViewData>(
     )
 
     private data class LinkSpanInfo(val text: String, val link: String)
-}
-
-/**
- * An [ArrayAdapter] that shows a "copy" button next to each item. When clicked
- * the text of the item is copied to the clipboard and a toast is shown (if
- * appropriate).
- */
-private class ArrayAdapterWithCopyButton<T : CharSequence>(
-    context: Context,
-    items: List<T>,
-) : ArrayAdapter<T>(context, R.layout.simple_list_item_1_copy_button, items) {
-    override fun getView(position: Int, convertView: View?, parent: ViewGroup): View {
-        val binding = if (convertView == null) {
-            SimpleListItem1CopyButtonBinding.inflate(LayoutInflater.from(context), parent, false)
-        } else {
-            SimpleListItem1CopyButtonBinding.bind(convertView)
-        }
-
-        getItem(position)?.let { text ->
-            binding.text1.text = text
-
-            binding.copy.setOnClickListener {
-                val clipboard = getSystemService(context, ClipboardManager::class.java) as ClipboardManager
-                val clip = ClipData.newPlainText("", text)
-                clipboard.setPrimaryClip(clip)
-                if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.S_V2) {
-                    Toast.makeText(
-                        context,
-                        context.getString(R.string.item_copied),
-                        Toast.LENGTH_SHORT,
-                    ).show()
-                }
-            }
-        }
-
-        return binding.root
-    }
 }
