+++ app/src/main/res/values-fi/strings.xml
@@ -765,6 +765,4 @@
     <string name="conversation_0_recipients">Ei muita osanottajia</string>
     <string name="compose_warn_language_dialog_accept_and_dont_ask_fmt">Julkaise muutoksetta (%1$s) äläkä kysy uudelleen</string>
     <string name="pref_title_confirm_status_language">Tarkasta julkaisun kieli ennen julkaisemista</string>
-    <string name="action_copy_item">Kopioi kohde</string>
-    <string name="item_copied">Teksti kopioitu</string>
 </resources>
