+++ core/ui/src/main/res/drawable/ic_content_copy_24.xml
