+++ core/ui/src/main/res/values-ga/strings.xml
@@ -14,4 +14,6 @@
     <string name="title_hashtags_dialog">Haischlibeanna</string>
     <string name="action_refresh">Athnuaigh</string>
     <string name="url_domain_notifier">" (🔗 %s)"</string>
+    <string name="item_copied">Cóipeáladh téacs</string>
+    <string name="action_copy_item">Cóipeáil mír</string>
 </resources>
