+++ core/ui/src/main/res/values/strings.xml
@@ -14,4 +14,6 @@
     <string name="action_hashtags">Hashtags</string>
     <string name="title_links_dialog">Links</string>
     <string name="title_hashtags_dialog">Hashtags</string>
+    <string name="item_copied">Text copied</string>
+    <string name="action_copy_item">Copy item</string>
 </resources>
